fix: properly sync search bar state with url (#167)
Router query is empty on the first render, which is when it is read, to set the initial state. To solve this document.location is used instead.

The component is rewritten as a functional component for consistency and ease of use.

The state for language is now the string 'en' or 'fr' instead of a boolean as well.